Further split out a player class which submits to "rooms"
